Rename files test_data* to data*
Files starting with 'test_' contains tests. Those files only contains data for tests.
