+++ magentoerpconnect/tests/test_sale_order.py
@@ -27,7 +27,7 @@ from .common import (mock_api,
                      mock_urlopen_image,
                      SetUpMagentoSynchronized,
                      )
-from .test_data import magento_base_responses
+from .data_base import magento_base_responses
 from .data_guest_order import guest_order_responses
 from ..sale import export_state_change
